initialize sb addon
