+++ .storybook/main.ts
@@ -1,2 +1,10 @@
-import main from '@lg-tools/storybook';
-export default main;
+export default {
+  addons: ['@lg-tools/storybook'],
+  framework: {
+    name: '@storybook/react-webpack5',
+    options: {
+      fastRefresh: true,
+      strictMode: true,
+    },
+  },
+};
