+++ package.json
@@ -50,7 +50,7 @@
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-router-dom": "5.2.0",
-    "storybook": "^7.1.0",
+    "storybook": "7.1.1",
     "turbo": "^1.10.7",
     "typescript": "~4.7.0"
   },
