--- tools/storybook/babel.config.js
@@ -1,2 +1,0 @@
-import babel from '@lg-tools/build/config/babel.config';
-export default babel;
