+++ tools/storybook/package.json
@@ -2,17 +2,19 @@
   "name": "@lg-tools/storybook",
   "version": "0.1.3",
   "description": "Storybook config for LeafyGreen",
-  "main": "main.ts",
   "license": "Apache-2.0",
+  "main": "./dist/index.js",
+  "module": "./dist/esm/index.js",
+  "types": "./dist/index.d.ts",
   "scripts": {
+    "build": "lg-internal-build-package",
+    "tsc": "tsc --build tsconfig.json",
     "upgrade": "npx storybook@latest upgrade --config-dir ."
   },
   "publishConfig": {
     "access": "public"
   },
   "dependencies": {
-    "@leafygreen-ui/palette": "^4.0.5",
-    "@leafygreen-ui/tokens": "^2.1.2",
     "@lg-tools/build": "^0.1.2",
     "@mdx-js/react": "1.6.22",
     "@storybook/addon-a11y": "7.1.1",
@@ -44,6 +46,11 @@
     "webpack": "5.88.0"
   },
   "peerDependencies": {
+    "@leafygreen-ui/lib": "^10.4.0",
+    "@leafygreen-ui/palette": "^4.0.5",
+    "@leafygreen-ui/tokens": "^2.1.2",
+    "@leafygreen-ui/typography": "^16.5.1",
+    "@lg-tools/storybook-decorators": "^0.1.1",
     "react": "^17.0.0",
     "react-dom": "^17.0.0",
     "storybook": "^7.1.0"
