+++ tools/storybook/src/main/utils.ts
@@ -1,0 +1,6 @@
+import isUndefined from 'lodash/isUndefined';
+import { RuleSetRule } from 'webpack';
+
+export function isRule(rule: any): rule is RuleSetRule {
+  return rule !== '...' && typeof rule === 'object' && !isUndefined(rule);
+}
